send init msg for uninit babylon contract
