+++ x/zoneconcierge/module_ibc.go
@@ -93,6 +93,10 @@ func (im IBCModule) OnChanOpenAck(
 	// if counterpartyVersion != types.Version {
 	// 	return errorsmod.Wrapf(types.ErrInvalidVersion, "invalid counterparty version: %s, expected %s", counterpartyVersion, types.Version)
 	// }
+
+	// this channel needs to be inited upon a newly finalised epoch
+	im.keeper.AddChannelToInit(ctx, channelID)
+
 	return nil
 }
 
@@ -102,6 +106,10 @@ func (im IBCModule) OnChanOpenConfirm(
 	portID,
 	channelID string,
 ) error {
+
+	// this channel needs to be inited upon a newly finalised epoch
+	im.keeper.AddChannelToInit(ctx, channelID)
+
 	return nil
 }
